+++ src/app/modules/+combat/combat.module.ts
@@ -1,9 +1,7 @@
 import { HttpClient } from "@angular/common/http";
 import { NgModule } from "@angular/core";
-import { HomeComponent } from "./home/home.component";
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
-import { HomeRoutingModule } from "./home-routing.module";
 import { MatInputModule } from "@angular/material/input";
 import { MatExpansionModule } from "@angular/material/expansion";
 import { MatChipsModule } from "@angular/material/chips";
@@ -12,6 +10,11 @@ import { MatButtonModule } from "@angular/material/button";
 import { MatDialogModule } from "@angular/material/dialog";
 import { MatTooltipModule } from "@angular/material/tooltip";
 import { CommonModule } from "@angular/common";
+import { CombatMainComponent } from "./combat-main/combat-main.component";
+import { CombatRoutingModule } from "./combat-routing.module";
+import { ReactiveFormsModule } from "@angular/forms";
+import { MatSelectModule } from '@angular/material/select';
+import { CombatDetailComponent } from './combat-detail/combat-detail.component';
 
 export const createTranslateLoader = (http: HttpClient) =>
   new TranslateHttpLoader(http, './assets/i18n/', '.json');
@@ -24,15 +27,18 @@ export const createTranslateLoader = (http: HttpClient) =>
     MatButtonModule,
     MatDialogModule,
     MatTooltipModule,
+    MatSelectModule,
   ];
 
 @NgModule({
   declarations: [
-    HomeComponent,
+    CombatMainComponent,
+    CombatDetailComponent,
   ],
   imports: [
     ...matModules,
-    HomeRoutingModule,
+    ReactiveFormsModule,
+    CombatRoutingModule,
     CommonModule,
     TranslateModule.forChild({
       loader: {
@@ -43,4 +49,4 @@ export const createTranslateLoader = (http: HttpClient) =>
     })
   ],
 })
-export class HomeModule { }
+export class CombatModule { }