+++ src/app/shared/components/menubar/menubar.component.ts
@@ -16,6 +16,10 @@ export class MenubarComponent {
     this.router.navigate(['characters']);
   }
 
+  public navigateCombat(){
+    this.router.navigate(['combat']);
+  }
+
   public navigateDetail(){
     //
   }