+++ src/app/shared/controllers/character-creation.controller.ts
@@ -1,3 +1,0 @@
-export const getUniqueID = (name: string) => {
-  return name + '_' + new Date().getTime();
-};