+++ src/app/shared/utils/character-creation.utils.ts
@@ -1,10 +1,34 @@
 import { TranslateService } from "@ngx-translate/core";
 import { Character, Location, Skill, hitpointsRatioEnum, weaponTypeEnum } from "../models/character.model";
 import { NUMBERS } from "../constants/number.constants";
-import { WeaponNameEnum, createWeapon } from "./weapon.factory";
+import { WeaponNameEnum, createWeapon } from "./equip.factory";
+import { getTotal } from "./dices.utils";
+
+export const resetTemporals = (pj: Character): void => {
+  pj.tempHPMod = NUMBERS.N_0;
+  pj.tempFPMod = NUMBERS.N_0;
+  pj.tempMPMod = NUMBERS.N_0;
+  pj.tempAPMod = NUMBERS.N_0;
+  pj.locations = pj.locations.map((loc: Location) => {
+    loc.bonusHP = NUMBERS.N_0;
+    loc.bonusAP = NUMBERS.N_0;
+    return loc;
+  });
+}
+export const setRandomHumanStats = (pj: Character): void => {
+  pj.stats.STR.value = getTotal(['3d6']);
+  pj.stats.DEX.value = getTotal(['3d6']);
+  pj.stats.CON.value = getTotal(['3d6']);
+  pj.stats.INT.value = getTotal(['2d6','+6']);
+  pj.stats.SIZ.value = getTotal(['2d6','+6']);
+  pj.stats.POW.value = getTotal(['3d6']);
+  pj.stats.CHA.value = getTotal(['3d6']);
+};
+
 
 export const setInitialHumanCharacter = (pj: Character, translate: TranslateService): Character => {
   const agiSkills =  translate.instant('PJ.SKILL_NAME.AGILITY');
+  pj.skills.AGILITY = [];
   pj.skills.AGILITY.push(new Skill(agiSkills.THROW, NUMBERS.N_25));
   pj.skills.AGILITY.push(new Skill(agiSkills.CLIMB, NUMBERS.N_40));
   pj.skills.AGILITY.push(new Skill(agiSkills.SWIM, NUMBERS.N_15));
@@ -15,6 +39,7 @@ export const setInitialHumanCharacter = (pj: Character, translate: TranslateServ
   pj.skills.AGILITY.push(new Skill(agiSkills.ROW, NUMBERS.N_5));
 
   const conSkills =  translate.instant('PJ.SKILL_NAME.COMUNICATION');
+  pj.skills.COMUNICATION = [];
   pj.skills.COMUNICATION.push(new Skill(conSkills.SPEAK_LANGUAGE, NUMBERS.N_30, translate.instant('PJ.SELF_LANGUAJE')));
   pj.skills.COMUNICATION.push(new Skill(conSkills.ACT, NUMBERS.N_0));
   pj.skills.COMUNICATION.push(new Skill(conSkills.SING, NUMBERS.N_5));
@@ -22,6 +47,7 @@ export const setInitialHumanCharacter = (pj: Character, translate: TranslateServ
   pj.skills.COMUNICATION.push(new Skill(conSkills.ORATORY, NUMBERS.N_5));
 
   const knowSkills =  translate.instant('PJ.SKILL_NAME.KNOWLEDGE');
+  pj.skills.KNOWLEDGE = [];
   pj.skills.KNOWLEDGE.push(new Skill(knowSkills.READ_WRITE, NUMBERS.N_0, translate.instant('PJ.SELF_LANGUAJE')));
   pj.skills.KNOWLEDGE.push(new Skill(knowSkills.APRAISE, NUMBERS.N_5));
   pj.skills.KNOWLEDGE.push(new Skill(knowSkills.CRAFT, NUMBERS.N_10));
@@ -37,6 +63,7 @@ export const setInitialHumanCharacter = (pj: Character, translate: TranslateServ
   pj.skills.KNOWLEDGE.push(new Skill(knowSkills.WORLD_LORE, NUMBERS.N_5));
 
   const perSkills =  translate.instant('PJ.SKILL_NAME.AWARENESS');
+  pj.skills.AWARENESS = [];
   pj.skills.AWARENESS.push(new Skill(perSkills.TOUCH, NUMBERS.N_10));
   pj.skills.AWARENESS.push(new Skill(perSkills.LISTEN, NUMBERS.N_25));
   pj.skills.AWARENESS.push(new Skill(perSkills.SCAN, NUMBERS.N_25));
@@ -46,6 +73,7 @@ export const setInitialHumanCharacter = (pj: Character, translate: TranslateServ
   pj.skills.AWARENESS.push(new Skill(perSkills.TASTE, NUMBERS.N_5));
 
   const manSkills =  translate.instant('PJ.SKILL_NAME.MANIPULATION');
+  pj.skills.MANIPULATION = [];
   pj.skills.MANIPULATION.push(new Skill(manSkills.CONCEAL, NUMBERS.N_5));
   pj.skills.MANIPULATION.push(new Skill(manSkills.INVENT, NUMBERS.N_5));
   pj.skills.MANIPULATION.push(new Skill(manSkills.DRIVE, NUMBERS.N_5));
@@ -53,10 +81,12 @@ export const setInitialHumanCharacter = (pj: Character, translate: TranslateServ
   pj.skills.MANIPULATION.push(new Skill(manSkills.PLAY_INSTRUMENT, NUMBERS.N_0));
 
   const steSkills =  translate.instant('PJ.SKILL_NAME.STEALTH');
+  pj.skills.STEALTH = [];
   pj.skills.STEALTH.push(new Skill(steSkills.HIDE, NUMBERS.N_10));
   pj.skills.STEALTH.push(new Skill(steSkills.MOVE_SILENTLY, NUMBERS.N_10));
 
   const magSkills =  translate.instant('PJ.SKILL_NAME.MAGICAL');
+  pj.skills.MAGICAL = [];
   pj.skills.MAGICAL.push(new Skill(magSkills.CEREMONY, NUMBERS.N_5));
   pj.skills.MAGICAL.push(new Skill(magSkills.ENCHANTMENT, NUMBERS.N_0));
   pj.skills.MAGICAL.push(new Skill(magSkills.INVOKE, NUMBERS.N_0));
@@ -66,10 +96,12 @@ export const setInitialHumanCharacter = (pj: Character, translate: TranslateServ
   pj.skills.MAGICAL.push(new Skill(magSkills.MULTISPELL, NUMBERS.N_0));
 
   const weapSkills =  translate.instant('PJ.SKILL_NAME.WEAPONS');
+  pj.skills.ATTACK = [];
   pj.skills.ATTACK.push(new Skill(weapSkills.NATURALS.FIST, NUMBERS.N_25));
   pj.skills.ATTACK.push(new Skill(weapSkills.NATURALS.KICK, NUMBERS.N_15));
   pj.skills.ATTACK.push(new Skill(weapSkills.MELEE.DAGGER, NUMBERS.N_15));
 
+  pj.skills.DEFENSE = [];
   pj.skills.DEFENSE.push(new Skill(weapSkills.NATURALS.FIST, NUMBERS.N_25));
   pj.skills.DEFENSE.push(new Skill(weapSkills.NATURALS.KICK, NUMBERS.N_15));
   pj.skills.DEFENSE.push(new Skill(weapSkills.MELEE.DAGGER, NUMBERS.N_15));
@@ -79,6 +111,7 @@ export const setInitialHumanCharacter = (pj: Character, translate: TranslateServ
   const fist = createWeapon(weaponTypeEnum.FIST, WeaponNameEnum.FIST, translate);
   const kick = createWeapon(weaponTypeEnum.KICK, WeaponNameEnum.KICK, translate);
 
+  pj.weapons = [];
   pj.weapons.push(fist);
   pj.weapons.push(kick);
 
@@ -87,6 +120,7 @@ export const setInitialHumanCharacter = (pj: Character, translate: TranslateServ
 
 const addHumanoidLocs = (pj: Character, translate: TranslateService) => {
   const locations =  translate.instant('PJ.LOCATION');
+  pj.locations = [];
   pj.locations.push(new Location(locations.RIGHT_LEG, hitpointsRatioEnum.X33));
   pj.locations.push(new Location(locations.LEFT_LEG, hitpointsRatioEnum.X33));
   pj.locations.push(new Location(locations.ABDOMEN, hitpointsRatioEnum.X33));
@@ -94,5 +128,9 @@ const addHumanoidLocs = (pj: Character, translate: TranslateService) => {
   pj.locations.push(new Location(locations.RIGHT_ARM, hitpointsRatioEnum.X25));
   pj.locations.push(new Location(locations.LEFT_ARM, hitpointsRatioEnum.X25));
   pj.locations.push(new Location(locations.HEAD, hitpointsRatioEnum.X33));
+};
+
+export const getUniqueID = (name: string) => {
+  return name + '_' + new Date().getTime();
 };